create new folder for every contacts controller and reexport them through index.js
