+++ controllers/index.js
@@ -1,0 +1,3 @@
+const contactsController = require("./contacts");
+
+module.exports = { contactsController };
